pyproject.toml
[tool.poetry]
name = "chairlift-api"
version = "0.1.0"
description = "Charlift API Server"
authors = ["BohdanSizomin <bohdansizomin@gmail.com>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.9"
fastapi = "^0.85.1"
uvicorn = "^0.19.0"
fastapi-pagination = "^0.10.0"
SQLAlchemy = "^1.4.42"
email-validator = "^1.3.0"
psycopg2-binary = "^2.9.4"
alembic = "^1.8.1"
invoke = "^1.7.3"
ipython = "^8.5.0"
sqladmin = "^0.6.1"
anyio = "^3.6.2"
setuptools = "^65.5.0"
python-dotenv = "^0.21.0"
passlib = "^1.7.4"
python-jose = "^3.3.0"
bcrypt = "^4.0.1"
fastapi-mail = "^1.2.0"
pytest-mock = "^3.10.0"

[tool.poetry.group.dev.dependencies]
flake8 = "^5.0.4"
black = "^22.10.0"
pytest = "^7.1.3"
autopep8 = "^1.7.0"
pytest-asyncio = "^0.20.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
